Test coordiantor indirectly
